requirements.txt
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
click==8.1.7
colorama==0.4.6
et-xmlfile==1.1.0
exceptiongroup==1.1.3
Flask==2.3.3
gunicorn==21.2.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
protobuf==4.21.12
pycparser==2.21
PySocks==1.7.1
python-dotenv==1.0.0
selenium==4.11.2
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.10.3
urllib3==2.0.4
Werkzeug==2.3.7
wsproto==1.2.0
